Minor Fixes
Added Shebang support.
There seems to be a bug in Espresso with the complete-match that
triggers the function completion, so it won't show up in the CodeSense
list. A bug report was filed and send to MacRabbit.
Also a minor bug was discovered, if you encounter CodeSense problems,
try adding a trailing new line. This seems to fix the problem (Thanks
@ianbech)